Simplify State/AccountsStore lifecycle by avoiding default->replace
